# docs/travel_api_swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Travel Planner API",
    "description": "Native PostgreSQL TravelerAPI with compare-and-swap locations and collection versioning.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4567",
      "description": "Development server"
    }
  ],
  "paths": {
    "/api/travel-plans": {
      "get": {
        "summary": "List travel plans",
        "description": "Retrieve all travel plans",
        "tags": [
          "Travel Plans"
        ],
        "responses": {
          "200": {
            "description": "List of travel plans",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TravelPlanSummary"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post": {
        "summary": "Create travel plan",
        "description": "Create a new travel plan",
        "tags": [
          "Travel Plans"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTravelPlanRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Travel plan created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TravelPlan"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/travel-plans/{id}": {
      "get": {
        "summary": "Get travel plan",
        "description": "Plan and ordered locations are read in one SQL statement and therefore share one READ COMMITTED snapshot.",
        "tags": [
          "Travel Plans"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/travel_plan_id"
          }
        ],
        "responses": {
          "200": {
            "description": "Travel plan details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TravelPlanDetails"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "put": {
        "summary": "Update travel plan",
        "description": "Update a travel plan using optimistic locking",
        "tags": [
          "Travel Plans"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/travel_plan_id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTravelPlanRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Travel plan updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TravelPlan"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/ConflictError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "delete": {
        "summary": "Delete travel plan",
        "description": "Delete a travel plan and all associated locations",
        "tags": [
          "Travel Plans"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/travel_plan_id"
          }
        ],
        "responses": {
          "204": {
            "description": "Travel plan deleted successfully"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/travel-plans/{id}/locations": {
      "post": {
        "summary": "Add location to plan",
        "description": "Add a new location to a travel plan (auto-ordered)",
        "tags": [
          "Locations"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/travel_plan_id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLocationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Location added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Location"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/locations/{id}": {
      "put": {
        "summary": "Update location",
        "description": "Partial location update. version is always required. visit_order additionally requires order_version; ordinary fields use location CAS without locking the parent plan.",
        "tags": [
          "Locations"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/location_id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLocationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Location updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Location"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "409": {
            "$ref": "#/components/responses/ConflictError"
          }
        }
      },
      "delete": {
        "summary": "Delete location",
        "description": "Delete a location from travel plan",
        "tags": [
          "Locations"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/location_id"
          }
        ],
        "responses": {
          "204": {
            "description": "Location deleted successfully"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "TravelPlan": {
        "type": "object",
        "required": [
          "id",
          "title",
          "currency",
          "is_public",
          "version",
          "created_at",
          "updated_at",
          "order_version"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the travel plan"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "Title of the travel plan"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Optional description of the travel plan"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "Start date of the trip"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "End date of the trip"
          },
          "budget": {
            "type": "number",
            "format": "decimal",
            "minimum": 0,
            "multipleOf": 0.01,
            "example": 150.75,
            "nullable": true,
            "description": "Budget amount with 2 decimal precision"
          },
          "currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$",
            "default": "USD",
            "description": "Currency code (ISO 4217)"
          },
          "is_public": {
            "type": "boolean",
            "default": false,
            "description": "Whether the plan is publicly visible"
          },
          "version": {
            "type": "integer",
            "minimum": 1,
            "description": "Version number for optimistic locking"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          },
          "order_version": {
            "type": "integer",
            "minimum": 1,
            "description": "Version of location membership and order. Changes on append, delete and reorder; ordinary field edits do not change it."
          }
        }
      },
      "TravelPlanSummary": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TravelPlan"
          },
          {
            "type": "object",
            "properties": {
              "location_count": {
                "type": "integer",
                "minimum": 0,
                "description": "Number of locations in this plan"
              }
            }
          }
        ]
      },
      "TravelPlanDetails": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TravelPlan"
          },
          {
            "type": "object",
            "properties": {
              "locations": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Location"
                },
                "description": "All locations in this travel plan, ordered by visit_order"
              }
            }
          }
        ]
      },
      "CreateTravelPlanRequest": {
        "type": "object",
        "required": [
          "title"
        ],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "Title of the travel plan"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Optional description"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "Start date of the trip"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "End date of the trip (must be >= startDate)"
          },
          "budget": {
            "type": "number",
            "minimum": 0,
            "nullable": true,
            "description": "Total budget for the trip"
          },
          "currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$",
            "default": "USD",
            "description": "Currency code (ISO 4217)"
          },
          "is_public": {
            "type": "boolean",
            "default": false,
            "description": "Whether the plan should be publicly visible"
          }
        }
      },
      "UpdateTravelPlanRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CreateTravelPlanRequest"
          },
          {
            "type": "object",
            "required": [
              "version"
            ],
            "properties": {
              "version": {
                "type": "integer",
                "minimum": 1,
                "description": "Current version number for optimistic locking"
              }
            }
          }
        ]
      },
...
      "CreateLocationRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "Name of the location"
          },
          "address": {
            "type": "string",
            "nullable": true,
            "description": "Address or description"
          },
          "latitude": {
            "type": "number",
            "minimum": -90,
            "maximum": 90,
            "nullable": true,
            "description": "Latitude coordinate"
          },
          "longitude": {
            "type": "number",
            "minimum": -180,
            "maximum": 180,
            "nullable": true,
            "description": "Longitude coordinate"
          },
          "arrival_date": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Planned arrival date and time"
          },
          "departure_date": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Planned departure date and time"
          },
          "budget": {
            "type": "number",
            "minimum": 0,
            "nullable": true,
            "description": "Budget allocated for this location"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "description": "Additional notes"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "ConflictResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Conflict error message"
          },
          "current_version": {
            "type": "integer",
            "description": "Current version of the resource"
          },
          "message": {
            "type": "string",
            "description": "Human-readable conflict explanation"
          },
          "current_order_version": {
            "type": "integer",
            "minimum": 1,
            "description": "Current collection version for an order conflict; replaces current_version in that case."
          }
        }
      }
    },
    "parameters": {
      "travel_plan_id": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Travel plan ID",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "location_id": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Location ID",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "responses": {
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "missingTitle": {
                "summary": "Missing required field",
                "value": {
                  "error": "Validation error",
                  "details": [
                    "Title is required"
                  ],
                  "timestamp": "2025-01-16T10:30:00Z"
                }
              },
              "invalid_date": {
                "summary": "Invalid date range",
                "value": {
                  "error": "Validation error",
                  "details": [
                    "End date must be after start date"
                  ],
                  "timestamp": "2025-01-16T10:30:00Z"
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "planNotFound": {
                "summary": "Travel plan not found",
                "value": {
                  "error": "Travel plan not found",
                  "timestamp": "2025-01-16T10:30:00Z"
                }
              },
              "locationNotFound": {
                "summary": "Location not found",
                "value": {
                  "error": "Location not found",
                  "timestamp": "2025-01-16T10:30:00Z"
                }
              }
            }
          }
        }
      },
      "ConflictError": {
        "description": "Optimistic locking conflict",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ConflictResponse"
            },
            "examples": {
              "version_conflict": {
                "summary": "Version conflict",
                "value": {
                  "error": "Conflict: Travel plan was modified by another user",
                  "currentVersion": 5,
                  "message": "Please refresh and try again"
                }
              }
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "serverError": {
                "summary": "Internal server error",
                "value": {
                  "error": "Internal server error",
                  "timestamp": "2025-01-16T10:30:00Z"
                }
              }
            }
          }
        }
      }
    }
  },
...
}